BIGC-637: added more error handling for delete item from cart, and added additional unit tests

+++ extension/lib/bigcommerce/CartRepository.js
@@ -145,6 +145,9 @@ class BigCommerceCartRepository {
    */
   async deleteProductFromCart (cartItemIds) {
     const cartId = await this._storage.get(CART_ID)
+    if (!cartId) {
+      throw (new Error('Cart ID is not available'))
+    }
     const deletePromises = []
     for (let cartItemId of cartItemIds) {
       deletePromises.push(this._client.delete('/carts/' + cartId + '/items/' + cartItemId))
